fix: Change speed type from float to double for improved precision in TrainEngine

+++ Scripts/Vanilla/TrainEngine.cs
@@ -3,9 +3,9 @@ namespace LandlessSkies.Vanilla;
 using Godot;
 
 public partial class TrainEngine : PathFollow3D {
-	[Export] public float speed = 3f;
+	[Export] public double speed = 3f;
 
 	public override void _Process(double delta) {
-		Progress += speed * (float)delta;
+		Progress += (float)(speed * delta);
 	}
 }
